Prevent object deletion on shutdown when checking !IsActive()
This is fairly ugly and sets an extension for the ConfigObjectUtility
delete handler to signal the OnActiveChanged handler inside the cluster
configsync to send a delete event to the other nodes.

refs #9927

+++ lib/remote/apilistener-configsync.cpp
@@ -53,7 +53,7 @@ void ApiListener::ConfigUpdateObjectHandler(const ConfigObject::Ptr& object, con
 	if (object->IsActive()) {
 		/* Sync object config */
 		listener->UpdateConfigObject(object, cookie);
-	} else {
+	} else if (!object->IsActive() && object->GetExtension("ConfigObjectDeleted")) {
 		/* Delete object */
 		listener->DeleteConfigObject(object, cookie);
 	}
